- refactor: replace some methods from MethodHelper to ReportHelper for report-related functionalities
